requirements.txt
requests>=2.28.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
Pillow>=10.0.0
